+++ plugin-api/src/jvmMain/kotlin/org/wip/plugintoolkit/api/processor/GeneratorUtils.kt
@@ -12,6 +12,7 @@ const val PLUGIN_ACTION_ANNOTATION = "org.wip.plugintoolkit.api.annotations.Plug
 const val RESUME_STATE_ANNOTATION = "org.wip.plugintoolkit.api.annotations.ResumeState"
 const val PLUGIN_SETUP_ANNOTATION = "org.wip.plugintoolkit.api.annotations.PluginSetup"
 const val PLUGIN_VALIDATE_ANNOTATION = "org.wip.plugintoolkit.api.annotations.PluginValidate"
+const val PLUGIN_LOAD_ANNOTATION = "org.wip.plugintoolkit.api.annotations.PluginLoad"
 
 object GeneratorUtils {
     fun mapKSTypeToDataType(ksType: KSType): DataType {